Create GenFence request
This change created real geofence request from the trigger, fence, and
target data. And the service handles updates when the fence is triggered
by creating a notification which can be use to open the MainActivity.java

+++ app/src/main/java/com/who/daola/data/Fence.java
@@ -9,7 +9,7 @@ public class Fence implements Serializable {
 
     private long mId;
     private String mName;
-    private double mRadius;
+    private float mRadius;
     private double mLatitude;
     private double mLongitude;
 
@@ -37,11 +37,11 @@ public class Fence implements Serializable {
         this.mLatitude = latitude;
     }
 
-    public double getRadius() {
+    public float getRadius() {
         return mRadius;
     }
 
-    public void setRadius(long radius) {
+    public void setRadius(float radius) {
         this.mRadius = radius;
     }
